renamed BaseLibrary.py to base_library.py, based on code review feedback
